# Reuse char[] buffer when decoding hex fields.