--- a/zipkin/src/main/java/zipkin2/internal/Proto3Fields.java
+++ b/zipkin/src/main/java/zipkin2/internal/Proto3Fields.java
@@ -164,6 +164,13 @@ final class Proto3Fields {
     static final char[] HEX_DIGITS =
       {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'};
 
+    // Reuse the buffer for decoding into hex since it's immediately copied into a String.
+    static final ThreadLocal<char[]> THIRTY_TWO_CHARS = new ThreadLocal<char[]>() {
+      @Override protected char[] initialValue() {
+        return new char[32];
+      }
+    };
+
     HexField(int key) {
       super(key);
     }
@@ -190,7 +197,13 @@ final class Proto3Fields {
 
     @Override String readValue(Buffer buffer, int length) {
       length *= 2;
-      char[] result = new char[length];
+
+      // All our hex fields are at most 32 characters.
+      if (length > 32) {
+        throw new AssertionError("hex field greater than 32 chars long: " + length);
+      }
+
+      char[] result = THIRTY_TWO_CHARS.get();
 
       for (int i = 0; i < length; i += 2) {
         byte b = buffer.readByte();
@@ -198,7 +211,7 @@ final class Proto3Fields {
         result[i + 1] = HEX_DIGITS[b & 0xf];
       }
 
-      return new String(result);
+      return new String(result, 0, length);
     }
   }