@adriancole 2019-05-11
PS brownie points for fixing `Span.Builder.traceId(long, long)` This is only used in thrift, but it has the same problem