@adriancole 2019-05-11
whoah I was wrong :) all the hex code in Span is used directly by brave...

ex.
```java
    Span.Builder builderWithContextData = Span.newBuilder()
        .traceId(context.traceIdString())
        .parentId(context.parentIdString())
        .id(context.spanIdString());
```

so I think if this is changed similarly, it will reduce allocations measurably on traced apps!