@adriancole 2019-05-11
PS I now remember why.. if you do an anonymous subclass, you can't unload the classloader. We have tests for this in Brave, but we don't here.