@adriancole 2019-05-11
this is the change that added tests in brave to ensure our code is unloadable.. I changed the threadlocal in b3 single to anonymous just to make sure and it failed the test https://github.com/apache/incubator-zipkin-brave/pull/787