@anuraaga 2019-05-11
Hmm - in that case does it make sense for me to add `zipkin2.internal.HexConversionBuffer` which can be used by brave too? No point to have multiple threadlocals all over.