@adriancole 2019-05-11
@anuraaga I think in the case of brave, probably duplicating one buffer is ok as it avoids osgi etc problems. We exclude internal packages right now. Maybe `Platform.idBuffer()` as hex is implied.