@anuraaga 2019-05-11
Got it - went ahead and extracted `Platform.idBuffer` in case we find other places in `zipkin` to use it, but won't use it directly in brave.