@anuraaga 2019-05-11
Ran benchmark again after merging - overall down from 2.2M to 1.7M

```
Benchmark                                                                    Mode    Cnt        Score        Error   Units
ProtoCodecBenchmarks.bytes_zipkinDecoder                                   sample  13762     1089.838 ±      5.981   us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.00         sample             945.152                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.50         sample            1039.360                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.90         sample            1224.704                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.95         sample            1359.872                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.99         sample            2033.664                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.999        sample            2733.335                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.9999       sample            9971.515                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p1.00         sample           12992.512                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate                    sample     15      973.403 ±     26.837  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate.norm               sample     15  1671164.718 ±     18.557    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space           sample     15      979.562 ±     68.569  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space.norm      sample     15  1682674.375 ± 124566.078    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen              sample     15        0.559 ±      0.565  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen.norm         sample     15      968.659 ±    991.452    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space       sample     15        0.666 ±      0.466  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space.norm  sample     15     1140.466 ±    790.469    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.count                         sample     15      102.000               counts
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.time                          sample     15       90.000                   ms
```